+++ server/post/manage/post-manage.template.tsx
@@ -1,6 +1,6 @@
 import React, { JSX } from "preact";
 import base from "server/base.template";
-import { ifTrue, ifSet, ifNotSet } from "server/macros/jsx-utils";
+import { ifTrue, ifSet, ifNotSet, ifFalse } from "server/macros/jsx-utils";
 import { CommonLocals } from "server/common.middleware";
 import links from "server/core/links";
 import * as formMacros from "server/macros/form.macros";
@@ -8,9 +8,10 @@ import { capitalize } from "lodash";
 import constants from "server/core/constants";
 import security from "server/core/security";
 import * as templatingFilters from "server/core/templating-filters";
+import { linkRestrictionAlert } from "server/user/components/link-restriction-alert.component";
 
 export default function render(context: CommonLocals): JSX.Element {
-  const { post, user, errorMessage, featuredEvent, specialPostType, allEvents, relatedEvent } = context;
+  const { post, user, errorMessage, featuredEvent, specialPostType, allEvents, relatedEvent, isTrustedUser } = context;
 
   formMacros.registerEditorScripts(context);
   formMacros.registerDatePickerScripts(context);
@@ -21,6 +22,8 @@ export default function render(context: CommonLocals): JSX.Element {
 
       <form method="post" class="js-warn-on-unsaved-changes">
         {context.csrfToken()}
+
+        {ifFalse(isTrustedUser, linkRestrictionAlert)}
 
         {ifTrue(errorMessage, () =>
           <div class="alert alert-warning">{errorMessage}</div>
@@ -77,7 +80,7 @@ export default function render(context: CommonLocals): JSX.Element {
 
         <div class="form-group">
           <label for="body">Body</label>
-          {formMacros.editor("body", post.get("body"), { autofocus: Boolean(post.get("id"))})}
+          {formMacros.editor("body", post.get("body"), { autofocus: Boolean(post.get("id")), noHyperlinks: !isTrustedUser })}
         </div>
 
         <div class="d-flex">
@@ -108,13 +111,13 @@ export default function render(context: CommonLocals): JSX.Element {
             <div class="form-group">
               <label for="published-at" class="mr-1">
                 Publication date
-            (in {templatingFilters.timezone(user.get("timezone")) || "UTC"}
+                (in {templatingFilters.timezone(user.get("timezone")) || "UTC"}
                 {ifNotSet(user.get("timezone"), () =>
                   <a href={links.routeUrl(user, "user", "settings")} class="btn btn-outline-secondary btn-sm">
                     <span class="fa fa-cog"></span>
                   </a>
                 )}
-            )
+                )
               </label>
               {formMacros.dateTimePicker("published-at", post.get("published_at"), user)}
             </div>