+++ server/post/post-view.controller.ts
@@ -4,6 +4,7 @@ import security from "server/core/security";
 import templating from "server/core/templating-functions";
 import { User } from "server/entity/user.entity";
 import { CustomRequest, CustomResponse } from "server/types";
+import userService from "server/user/user.service";
 import eventService from "../event/event.service";
 import commentService from "./comment/comment.service";
 import likeService from "./like/like.service";
@@ -38,11 +39,12 @@ export async function postView(req: CustomRequest, res: CustomResponse<PostLocal
  */
 export async function buildPostContext(locals: PostLocals): Promise<PostLocals> {
   const post: PostBookshelfModel = locals.post;
-  const currentUser: User = locals.user;
+  const currentUser: User | undefined = locals.user;
 
   const context = {
     ...locals,
     allEvents: (await eventService.findEvents()).models,
+    isTrusterUser: currentUser ? (await userService.isTrustedUser(locals.user)) : false,
     relatedEvent: undefined,
     relatedEntry: undefined,
     specialPostType: undefined,